Hotfix: Fix postgres setup test db if the db doesn't exist already
This fixes the postgres setup test db to handle if the db doesn't
exist already. This was missed when originally developing the
setup test db file because the db was created at some point and I
forgot to go back and remove it and try running the test suite
again. This fixes it by not establishing a connection until after
the db has been dropped and re-created. This allows the test suite
to run if the db doesn't exist.