Pass `Config` to `Migration` instead of using `Ardb.config`
This switches to passing a `Config` to the `Migration` class
isntead of it referencing the global `Ardb.config`. This is to
decouple it and make it so its tests don't require ardb to be
configured and init. This doesn't change the migration class's
functionality.

This also updates the cli generate migration command to use
`Ardb.config` when it builds a `Migration`. This keeps the cli
working as it previously did.