cli: update valid commands to allow custom CLIRB build procs
This allows individual commands to take individual options.  There
is no immediate need for this behavior but it was added to GGem (see
GGem's PR 35).  Ardb's command handling is modeled after GGem so this
is added to keep things consistent.

Note: this also updates to the latest Assert patch.  This is just
to keep up with the latest stuff.