Add nice error when ardb can't require a db file on init
This updates ardb to throw a nice error when it can't require the
db file specified by the `ARDB_DB_FILE` env var. This is to help
users out when using hte CLI or ardb in their code.

This also updates the cli commands to specially handle this error
by moving `Ardb.init` out of their begin rescue blocks. This makes
it so the error doesn't get caught by the commands custom error
handling. Each command does custom error handling to better explain
that the action they were trying to perform couldn't be done. The
invalid db file error would be stomped by this error handling so
users wouldn't get the nice error message. By moving it out, the
inavlid db file error is handled by the CLI which displays it to
the user. Any errors thrown by `Ardb.init` should be handled this
way because they aren't directly related to the action being
performed and they have error messages that are designed to help
users.

Closes #100