Update `escape_like_pattern` to always return the pattern string
This updates the `escape_like_pattern` method on the adapters to
always ensure it returns the pattern that was being escaped. The
`gsub!` method will only return the string if it matched, otherwise
it returns `nil`. This caused tools using `escape_like_pattern` to
not work if nothing needed to be escaped. This was missed because
the tests only tested how it behaved with a string that the gsubs
matched.