Raise not initialized error when using `adapter` before `init`
This updates ardb to raise a not initialized error when its adapter
is used before it has been `init`. This is to avoid no method
errors happening because the `adapter` is `nil`. Instead this
throws a nice error that explains ardb hasn't been init yet.

This also adds `reset_adapter` method which allows setting the
adapter back to `nil`. This was needed to reliably test that it
would throw the not initialized error. Ardb has system tests that
init ardb so its adapter can be set depending on the order the
tests are run. This updates the tests to reset the adapter to
help ensure we aren't depending on the global state in our tests.

This also updates a couple locations that change exceptions
backtraces to avoid them pointing to the internals of ardb. They
previously used `tap` to one-line this operation but blocks change
ruby's `caller` so the backtrace wasn't the caller of the method
but the caller of `tap` (which pointed to ardb internals). This
switches to setting the backtrace on its own line outside of a
block to ensure `caller` is the methods caller. I noticed this
while testing this with a real use-case of ardb not being init
before using adapter and saw the wrong backtrace.

Closes #116