validate the configured database name has no non-word characters
This updates the validation logic to add this validation. I ran
into problems running commands on databases with non-word
characters in their name. Specifically, I couldn't drop a postgres
db with non-word characters in the name.

This validation prevents/protects the user from getting in a bad
state b/c of the database name.