update commands for better UX around the database existing
The updates the CLI commands so that they detect whether the
database exists or not and output friendly error messages if the
database is in a bad state for the command being attempted:

```
$ ardb connect
error: database "some_database" does not exist.
$ ardb drop
error: database "some_database" does not exist.
$ ardb migrate
error: database "some_database" does not exist.
$ ardb create
created postgresql db `some_database`
$ ardb create
error: database "some_database" already exists.
```

Previously, the commands would spit out raw exception style
output with raw attempted SQL and/or backtraces. This isn't
really friendly or necessary when the root cause of the problem
is this obvious.