rework commands to configure their own name/summary
This makes it so you just add commands to the command set and
don't have to specify the command name.

After adding the new migration commands, it felt weird to configure
the summary and help message on the command object but configure
the commmand name separately. Also the command name in the help
message had to be kept in sync with the configured command name.

This reworks so that the command name is configured like the
summary and the help message is now a generated template on the
ValidCommand mix-in.

This also adds tests for the new migration commands which I forgot
to add in the previous commit where I added the commands
themselves.