cleanup CLI output based on writing the README
The README includes a bunch of CLI output to give an impression
of what working with the CLI is like and what tools it offers.
While preparing the README, I noticed some inconsistencies with
the output on success vs error on certain commands. This cleans
up the output to always use inspect instead of back-ticks.

I also noticed that the help output for `generate-migration`
didn't describe the MIGRATION-NAME parameter. This was an error
from previously reworking the CLI logic. I correct this here by
overriding the help message to describe the MIGRATION-NAME
parameter.