add a first-pass README
This adds some description and usage information. This really just
covers configuration and the CLI. I chose to not highlight some of
the other testing related tools as I think they need to be updated
for Rails 5. Others are tied to using Assert and are esoteric.

As I address the other tools, I'll add README descriptions for
them later on.