Update "require autoloaded" script to support ActiveRecord 6.1
This updates the `require_autoloaded_active_record_files` script
to support ActiveRecord 6.1. This was previously updated to
support ActiveRecord 6.0 but some files have been removed and
added in ActiveRecord 6.1 and the previous didn't work for it.

This also fixes an error that would occur when
`ActiveRecord::MigrationContext` wasn't required before running
the `Adapter::Base` tests. This requires `active_record/migration`
to ensure it's always available.

Finally, this updates to support ruby 2.5 and greater going
forward to match other gems.